Fixed many warnings.

+++ src/library/entity/MeshEntity.cpp
@@ -94,7 +94,7 @@ namespace PR
 	FacePoint MeshEntity::getRandomFacePoint(Sampler& sampler, Random& random) const
 	{
 		auto ret = sampler.generate3D();
-		uint32 fi = Projection::map(PM::pm_GetX(ret), 0, mMesh->faces().size() - 1);
+		uint32 fi = Projection::map(PM::pm_GetX(ret), 0, (int)mMesh->faces().size() - 1);
 
 		Face* face = mMesh->getFace(fi);
